03_pictionary/requirements.txt
python-fasthtml
uvicorn>=0.29
mistralai